Refactor sync/consensus graph test related code (#420)
