+++ core/src/consensus/anticone_cache.rs
@@ -1,3 +1,7 @@
+// Copyright 2019 Conflux Foundation. All rights reserved.
+// Conflux is free software and distributed under GNU General Public License.
+// See http://www.gnu.org/licenses/
+
 use hibitset::{BitSet, BitSetLike};
 use std::{
     cmp::max,
