+++ core/src/consensus/consensus_inner/consensus_new_block_handler.rs
@@ -1,3 +1,7 @@
+// Copyright 2019 Conflux Foundation. All rights reserved.
+// Conflux is free software and distributed under GNU General Public License.
+// See http://www.gnu.org/licenses/
+
 use crate::{
     block_data_manager::{BlockDataManager, BlockStatus, LocalBlockInfo},
     consensus::{
