+++ core/src/consensus/consensus_inner/mod.rs
@@ -1,3 +1,7 @@
+// Copyright 2019 Conflux Foundation. All rights reserved.
+// Conflux is free software and distributed under GNU General Public License.
+// See http://www.gnu.org/licenses/
+
 pub mod consensus_executor;
 pub mod consensus_new_block_handler;
